[clojure.test] Use regular :fail instead of custom :mismatch message
In case of a mismatch we emit a custom map with {:type :mismatch} in order to do
custom rendering. The problem with this approach is that other tools and
reporters (CIDER, Cursive, eftest, ...) all need to be aware of :mismatch.

Instead emit a standard {:type :fail} method, but extend
`clojure.core/print-method` to still get the pretty printing of mismatch
information. This does mean we are forced in the straightjacket of "Expected:
...\nActual: ...".